refactor(analytics): reduce the schema validator to a metric-key parser
The validator existed to probe metric_catalog rows against ClickHouse and
write their schema_status back. With that table unread, only its metric-key
parser still has a consumer: metric-results request validation.

Keeps domain/metric_key.rs (moved from schema_validator/parse.rs) and drops
the rest of the module along with AppState.validator and the startup sweep.
The metric-definition validator is unrelated and stays.
