+++ src/backend/services/analytics/src/api/mod.rs
@@ -29,7 +29,6 @@ use utoipa::openapi::schema::{
 use crate::config::GearConfig;
 use crate::domain::metric_definitions::listing as metric_definitions_listing;
 use crate::domain::saved_query;
-use crate::domain::schema_validator::SchemaValidator;
 use crate::infra::identity::IdentityClient;
 
 /// Shared application state.
@@ -40,11 +39,6 @@ pub struct AppState {
     pub identity: IdentityClient,
     #[allow(dead_code)] // will be used for runtime config access (rate limits, feature flags)
     pub config: GearConfig,
-    /// Schema-validator (Refs #521). Retained so a future per-write path can
-    /// call `validator.validate(metric_key)`; the startup sweep owns its own
-    /// clone.
-    #[allow(dead_code)] // the startup sweep holds its own clone
-    pub validator: SchemaValidator,
 }
 
 pub(crate) fn forwarded_authorization(headers: &axum::http::HeaderMap) -> Option<&str> {
